refactor: 테스트 엔티티 팩토리를 bottlenote-mono로 통합
## 변경 사항

### 구조 개선
- 분산된 테스트 엔티티 팩토리를 중앙 집중 관리
- bottlenote-mono/src/test/java/app/bottlenote/fixture/ 로 통합
- 엔티티와 테스트 픽스처의 논리적 응집도 향상

### 이동된 파일 (4개)
- UserTestFactory.java
- AlcoholTestFactory.java
- RatingTestFactory.java
- BusinessSupportTestFactory.java

### 패키지 경로 변경
- Before: app.bottlenote.{domain}.fixture.*TestFactory
- After: app.bottlenote.fixture.*TestFactory

### 영향받은 테스트 (3개)
- UserMyPageControllerTest
- AlcoholQueryIntegrationTest
- BusinessSupportIntegrationTest

## 목적
- 모든 테스트 엔티티 팩토리를 한눈에 파악 가능
- 일관된 작성 규칙 적용 기반 마련
- 엔티티(bottlenote-mono/domain)와 같은 모듈에 배치하여 응집도 향상
