+++ bottlenote-mono/src/test/java/app/bottlenote/fixture/UserTestFactory.java
@@ -1,4 +1,4 @@
-package app.bottlenote.user.fixture;
+package app.bottlenote.fixture;
 
 import app.bottlenote.user.constant.FollowStatus;
 import app.bottlenote.user.constant.GenderType;
