+++ bottlenote-product-api/src/test/java/app/bottlenote/alcohols/integration/AlcoholQueryIntegrationTest.java
@@ -15,10 +15,10 @@ import app.bottlenote.alcohols.domain.TastingTag;
 import app.bottlenote.alcohols.dto.response.AlcoholDetailResponse;
 import app.bottlenote.alcohols.dto.response.AlcoholSearchResponse;
 import app.bottlenote.alcohols.dto.response.AlcoholsSearchItem;
-import app.bottlenote.alcohols.fixture.AlcoholTestFactory;
-import app.bottlenote.rating.fixture.RatingTestFactory;
+import app.bottlenote.fixture.AlcoholTestFactory;
+import app.bottlenote.fixture.RatingTestFactory;
+import app.bottlenote.fixture.UserTestFactory;
 import app.bottlenote.user.domain.User;
-import app.bottlenote.user.fixture.UserTestFactory;
 import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
